Massage `BaseContainer.TryFindBlockAt` method for 2x performance for single-block case.
Very subtle things caused slowdown, e.g. Get<> instead of GetRef<>.

Also rename chunk to block. We do have a chain of them and chunk chain is so weird, go for the hype FTW!

+++ src/Spreads.Core/Lookup.cs
@@ -55,7 +55,8 @@ namespace Spreads
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public static bool IsEqualityOK(this Lookup lookup)
         {
-            return ((int)lookup & (int)(Lookup.EQ)) != 0;
+            // ReSharper disable once BitwiseOperatorOnEnumWithoutFlags
+            return (lookup & Lookup.EQ) != 0;
         }
 
         /// <summary>
